+++ tests/test_converters.py
@@ -1,4 +1,8 @@
 from vecorel_cli.converters import Converters
+from vecorel_cli.vecorel.schemas import VecorelSchema
+from vecorel_cli.vecorel.util import load_file
+
+from fiboa_cli.fiboa.version import get_fiboa_uri
 
 
 def test_converters(capsys):
@@ -21,3 +25,21 @@ def test_changed_properties():
         converter = c.load(_id)
         assert getattr(converter, "license") is None or isinstance(converter.license, str)
         assert getattr(converter, "provider") is None or isinstance(converter.provider, str)
+
+
+def test_overriden_base_properties():
+    """
+    You should not define a different schema for a property if it is defined in the base schema.
+    """
+    c = Converters()
+    for _id in Converters().list_ids():
+        converter = c.load(_id)
+        schemas = converter.missing_schemas
+        converter_properties = schemas and schemas.get("properties") or {}
+        schema = VecorelSchema(load_file(get_fiboa_uri()))
+
+        for property, s in schema["properties"].items():
+            if property in converter_properties:
+                assert s == converter_properties[property], (
+                    "Converter {converter} overrides schema for base property {property}"
+                )